pyproject.toml
[project]
name = "flash-tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "loguru>=0.7.3",
    "markdown>=3.10.1",
    "pygments>=2.19.2",
    "pyyaml>=6.0.3",
    "google-generativeai",
    "openai",
]
